requirements.txt
asgiref==3.3.1
Django==3.1.7
django-dbbackup==3.3.0
django-environ==0.4.5
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
